requirements.txt
streamlit
pandas
matplotlib
nltk
python-dotenv
google-generativeai
